Add authorization using jsonwebtokens and local storage. And add a 'user' property to the workout model

+++ frontend/src/components/WorkoutDetails.js
@@ -1,18 +1,26 @@
-import React from 'react'
+import React, { useContext } from 'react'
 import useWorkoutsContext from '../hooks/useWorkoutsContext'
 
 //Date format
 import {formatDistanceToNow} from "date-fns"
+import { AuthContext } from '../context/AuthContext'
 
 function WorkoutDetails(props) {
     const {title, reps, load, createdAt, _id} = props.workout
     const {dispatch} = useWorkoutsContext()
+    const {state} = useContext(AuthContext)
 
     // Delete request to erase from Database
     const handleDelete = async ()=>{
+        if(!state.token) {
+            return
+        }
         const response = await fetch("http://localhost:4000/api/workouts/" + _id, {
             method: "DELETE",
-            headers: {"Content-Type": "application/json"}
+            headers: {
+                "Content-Type": "application/json",
+                "Authorization": `Bearer ${state.token}`
+            }
         })
         if(response.ok){
             dispatch({type: "DELETE_WORKOUT", payload: _id})
